+++ cocos2dx/platform/mac/EAGLView.mm
@@ -329,7 +329,7 @@ static EAGLView *view;
 	float x = local_point.x;
 	float y = [self getHeight] - local_point.y;
 	
-    int ids[1] = {0};
+    long ids[1] = {0};
     float xs[1] = {0.0f};
     float ys[1] = {0.0f};
     
@@ -353,7 +353,7 @@ static EAGLView *view;
 	float x = local_point.x;
 	float y = [self getHeight] - local_point.y;
 
-    int ids[1] = {0};
+    long ids[1] = {0};
     float xs[1] = {0.0f};
     float ys[1] = {0.0f};
     
@@ -372,7 +372,7 @@ static EAGLView *view;
 	float x = local_point.x;
 	float y = [self getHeight] - local_point.y;
 
-    int ids[1] = {0};
+    long ids[1] = {0};
     float xs[1] = {0.0f};
     float ys[1] = {0.0f};
